L4 plankton abundance time-series

This dataset is a plankton abundance time-series collected at the Western Channel Observatory L4 station. It contains zooplankton abundances in #/m^3, phytoplankton abundances in #/ml, and sea surface temperature in degrees C. Sampling dates for zooplankton, phytoplankton, and surface temperature values are not always aligned, resulting in NA values within the time-series.

Usage

data(L4.AllDates)

Format

A data frame with 1123 observations on the following 20 variables.

date	sampling dates; often read into R as a factor variable

cnidarian	e.g., hydromedusae and siphonophores

amphipod	e.g., gammarids

chaetognath	Sagitta spp.

krill	Euphausiids (all stages)

pteropod	Clione and Limacina

tunicate	larvaceans

cladoceran	Evadne spp. and Podon spp.

calanoid.lg	large calanoid copepods (>2 mm)

calanoid.sm	small calanoid copepods (<2 mm)

cyclopoid	cyclopoid copepods

poecilostom	poecilostomatoid copepods

harpact	harpacticoid copepods

diatom	phytoplankton: diatoms

dino	phytoplankton: dinoflagellates

other.algae	phytoplankton: e.g., coccolithophores, green

cirripedia	meroplankton: barnacle nauplii and cyprids

mero.grazers	meroplankton: other grazers, e.g., gastropod and bivalve larvae

decapod	meroplankton: e.g., crab and shrimp larvae

surface.temp	sea surface temperature (degrees C)

Source

http://www.westernchannelobservatory.org.uk/data.php

References

Eloire D, Somerfield PJ, Conway DVP, Halsband-Lenk C, Harris R, and Bonnet D (2010) Temporal variability and community composition of zooplankton at station L4 in the Western Channel: 20 years of sampling. Journal of Plankton Research 32:657-679 [doi:10.1093/plankt/fbq009]

Southward AJ, Langmead O, Hardman-Mountford NJ, Aiken J, Boalch GT, Dando PR, Genner MJ, Joint I, Kendall MA, Halliday NC, Harris RP, Leaper R, Mieszkowska N, Pingree RD, Richardson AJ, Sims DW, Smith T, Walne AW, and Hawkins SJ (2005) Long-term oceanographic and ecological research in the western English Channel. Advances in Marine Biology 47:1-105 [doi:10.1016/S0065-2881(04)47001-1]

Scheef, LP, DE Pendleton, SE Hampton, SL Katz, EE Holmes, ME Scheuerell, and DG Johns (2012) Assessing marine plankton community structure from long-term monitoring data with multivariate autoregressive (MAR) models: a comparison of fixed station vs. spatially distributed sampling data. Limnology & Oceanography: Methods 10:54-64.

See also

Examples

data(L4.AllDates)
summary(L4.AllDates)
#>         date        cnidarian          amphipod        chaetognath    
#>  1/10/1996:   1   Min.   :   0.00   Min.   :  0.000   Min.   :  0.00  
#>  1/10/2000:   1   1st Qu.:   3.60   1st Qu.:  0.000   1st Qu.:  3.15  
#>  1/11/1999:   1   Median :  25.53   Median :  0.000   Median : 13.12  
#>  1/12/1998:   1   Mean   :  85.21   Mean   :  2.501   Mean   : 30.96  
#>  1/13/1992:   1   3rd Qu.:  93.87   3rd Qu.:  1.462   3rd Qu.: 35.43  
#>  1/13/1997:   1   Max.   :4798.54   Max.   :103.230   Max.   :568.34  
#>  (Other)  :1117   NA's   :220       NA's   :220       NA's   :220     
#>      krill           pteropod          tunicate          cladoceran    
#>  Min.   : 0.000   Min.   :   0.00   Min.   :   0.000   Min.   :   0.0  
#>  1st Qu.: 0.000   1st Qu.:   0.00   1st Qu.:   1.595   1st Qu.:   0.0  
#>  Median : 0.000   Median :   4.18   Median :  29.400   Median :  19.6  
#>  Mean   : 1.166   Mean   :  31.17   Mean   :  93.435   Mean   : 236.9  
#>  3rd Qu.: 0.320   3rd Qu.:  17.54   3rd Qu.: 106.710   3rd Qu.: 192.7  
#>  Max.   :40.160   Max.   :3435.78   Max.   :3159.970   Max.   :7634.1  
#>  NA's   :220      NA's   :220       NA's   :220        NA's   :220     
#>   calanoid.lg       calanoid.sm         cyclopoid       poecilostom     
#>  Min.   :   0.00   Min.   :    1.97   Min.   :   0.0   Min.   :   0.00  
#>  1st Qu.:  10.86   1st Qu.:  471.01   1st Qu.: 100.4   1st Qu.:  65.58  
#>  Median :  36.86   Median :  901.52   Median : 208.9   Median : 194.75  
#>  Mean   :  78.01   Mean   : 1292.39   Mean   : 363.6   Mean   : 412.37  
#>  3rd Qu.:  91.55   3rd Qu.: 1625.71   3rd Qu.: 425.5   3rd Qu.: 548.02  
#>  Max.   :1915.19   Max.   :11017.49   Max.   :4790.6   Max.   :5745.03  
#>  NA's   :220       NA's   :220        NA's   :220      NA's   :220      
#>     harpact           diatom             dino           other.algae     
#>  Min.   :  0.00   Min.   :   0.36   Min.   :   0.000   Min.   :  279.7  
#>  1st Qu.:  4.73   1st Qu.:   6.92   1st Qu.:   0.102   1st Qu.: 1563.0  
#>  Median : 16.37   Median :  25.74   Median :   0.740   Median : 2359.0  
#>  Mean   : 41.36   Mean   : 166.75   Mean   :  68.933   Mean   : 2811.1  
#>  3rd Qu.: 46.95   3rd Qu.: 119.15   3rd Qu.:   9.957   3rd Qu.: 3655.1  
#>  Max.   :818.95   Max.   :5055.86   Max.   :4810.810   Max.   :10882.0  
#>  NA's   :220      NA's   :496       NA's   :496        NA's   :496      
#>    cirripedia        mero.grazers        decapod         surface.temp  
#>  Min.   :    0.00   Min.   :   0.00   Min.   :  0.000   Min.   : 7.30  
#>  1st Qu.:    0.00   1st Qu.:  25.01   1st Qu.:  3.155   1st Qu.:10.20  
#>  Median :   12.63   Median :  64.26   Median :  9.980   Median :12.80  
#>  Mean   :  321.14   Mean   : 169.01   Mean   : 15.946   Mean   :12.82  
#>  3rd Qu.:   70.07   3rd Qu.: 170.24   3rd Qu.: 21.495   3rd Qu.:15.35  
#>  Max.   :19543.52   Max.   :6780.94   Max.   :160.000   Max.   :20.30  
#>  NA's   :220        NA's   :220       NA's   :220       NA's   :245