L4 plankton abundance time-series prepared for MAR modeling

Plankton abundance data collected at the Western Channel Observatory L4 station, formatted for use in run.mar. The dataset was prepared for MAR analysis with a call to
prepare.data(data=L4.AllDates, increment="month", fill.gap=0, replace.0s="rand.half",
log=T, z.method="deseason")

Usage

data(L4.mar)

Format

A data frame with 179 observations on the following 21 variables.

contin	a numeric vector where matching values indicate blocks of continuous sampling dates

date	dates

cnidarian	e.g., hydromedusae and siphonophores

amphipod	e.g., gammarids

chaetognath	Sagitta spp.

krill	Euphausiids (all stages)

pteropod	Clione and Limacina

tunicate	larvaceans

cladoceran	Evadne spp. and Podon spp.

calanoid.lg	large calanoid copepods (>2 mm)

calanoid.sm	small calanoid copepods (<2 mm)

cyclopoid	cyclopoid copepods

poecilostom	poecilostomatoid copepods

harpact	harpacticoid copepods

diatom	phytoplankton: diatoms

dino	phytoplankton: dinoflagellates

other.algae	phytoplankton: e.g., coccolithophores, green

cirripedia	meroplankton: barnacle nauplii and cyprids

mero.grazers	meroplankton: other grazers, e.g., gastropod and bivalve larvae

decapod	meroplankton: e.g., crab and shrimp larvae

surface.temp	sea surface temperature

Source

http://www.westernchannelobservatory.org.uk/data.php

See also

Examples

data(L4.mar)
summary(L4.mar)
#>      contin           date              cnidarian          amphipod      
#>  Min.   :56.00   Min.   :1992-10-16   Min.   :-3.2865   Min.   :-2.4689  
#>  1st Qu.:64.00   1st Qu.:1997-03-01   1st Qu.:-0.5210   1st Qu.:-0.7326  
#>  Median :65.00   Median :2000-12-16   Median : 0.1697   Median : 0.2740  
#>  Mean   :63.47   Mean   :2000-10-29   Mean   : 0.0000   Mean   : 0.0000  
#>  3rd Qu.:65.00   3rd Qu.:2004-08-31   3rd Qu.: 0.7000   3rd Qu.: 0.6708  
#>  Max.   :66.00   Max.   :2008-06-15   Max.   : 2.4775   Max.   : 2.1308  
#>   chaetognath          krill            pteropod          tunicate      
#>  Min.   :-3.1162   Min.   :-2.6978   Min.   :-2.8906   Min.   :-2.9482  
#>  1st Qu.:-0.5270   1st Qu.:-0.8982   1st Qu.:-0.4429   1st Qu.:-0.6504  
#>  Median : 0.1316   Median : 0.1566   Median : 0.2653   Median : 0.1040  
#>  Mean   : 0.0000   Mean   : 0.0000   Mean   : 0.0000   Mean   : 0.0000  
#>  3rd Qu.: 0.7415   3rd Qu.: 0.8896   3rd Qu.: 0.7080   3rd Qu.: 0.6717  
#>  Max.   : 1.7663   Max.   : 1.7534   Max.   : 1.6129   Max.   : 2.1902  
#>    cladoceran       calanoid.lg        calanoid.sm        cyclopoid       
#>  Min.   :-3.1569   Min.   :-3.02708   Min.   :-3.1560   Min.   :-3.28323  
#>  1st Qu.:-0.5652   1st Qu.:-0.49160   1st Qu.:-0.6142   1st Qu.:-0.51031  
#>  Median : 0.1191   Median : 0.07195   Median : 0.0849   Median :-0.06296  
#>  Mean   : 0.0000   Mean   : 0.00000   Mean   : 0.0000   Mean   : 0.00000  
#>  3rd Qu.: 0.6927   3rd Qu.: 0.65477   3rd Qu.: 0.6177   3rd Qu.: 0.60486  
#>  Max.   : 2.1568   Max.   : 2.24018   Max.   : 2.3226   Max.   : 2.16504  
#>   poecilostom         harpact            diatom              dino         
#>  Min.   :-2.4402   Min.   :-3.1163   Min.   :-2.46989   Min.   :-1.96857  
#>  1st Qu.:-0.6176   1st Qu.:-0.4528   1st Qu.:-0.65844   1st Qu.:-0.64685  
#>  Median : 0.0183   Median : 0.2257   Median :-0.09897   Median :-0.09827  
#>  Mean   : 0.0000   Mean   : 0.0000   Mean   : 0.00000   Mean   : 0.00000  
#>  3rd Qu.: 0.6257   3rd Qu.: 0.6748   3rd Qu.: 0.69180   3rd Qu.: 0.59230  
#>  Max.   : 2.9271   Max.   : 1.8738   Max.   : 2.98016   Max.   : 2.73673  
#>   other.algae        cirripedia       mero.grazers         decapod       
#>  Min.   :-2.6888   Min.   :-3.1870   Min.   :-2.97017   Min.   :-3.3886  
#>  1st Qu.:-0.6996   1st Qu.:-0.5470   1st Qu.:-0.56894   1st Qu.:-0.4662  
#>  Median :-0.0427   Median : 0.1377   Median : 0.03999   Median : 0.2038  
#>  Mean   : 0.0000   Mean   : 0.0000   Mean   : 0.00000   Mean   : 0.0000  
#>  3rd Qu.: 0.7258   3rd Qu.: 0.6730   3rd Qu.: 0.58079   3rd Qu.: 0.5349  
#>  Max.   : 2.1293   Max.   : 2.7458   Max.   : 2.25556   Max.   : 2.5506  
#>   surface.temp    
#>  Min.   :-1.9121  
#>  1st Qu.:-0.7550  
#>  Median :-0.0584  
#>  Mean   : 0.0000  
#>  3rd Qu.: 0.6639  
#>  Max.   : 2.4538