Plot coefficients of a MAR model

Plots the coefficients contained in an object of class MAR resulting from a call to run.mar

# S3 method for MAR
plot(x, y=NULL, ..., legend=FALSE)

Arguments

x, y, ...

Objects of class MAR to be plotted

legend

Should a legend be generated? If set to TRUE, an additional graphics device containing the legend will be generated

Details

The B- and C-matrix coefficients of the best-fit model are plotted as a grid of bar graphs. If bootstrapping was performed, discarded best-fit coefficients are faded. If any restrictions were set on interactions, red and green points at the base of bars are indicative of exclusion and inclusion, respectively.

Value

A graphics device containing the coefficient plot, and, if legend=TRUE, an additional device containing the corresponding legend

Author

LP Scheef